Изменил систему работы с grunt'ом установил пакет load-grunt-config, который упрощает работу с различными тасками
