+++ grunt/aliases.yaml
@@ -1,0 +1,17 @@
+default:
+	description: 'Сборка по умолчанию (разработка)'
+	tasks:
+		- dev
+
+
+dev:
+	description: 'Сборка для разработки'
+	tasks:
+		- 'concurrent:target1'
+		- 'concurrent:target2'
+
+task:
+	description: 'Запуск всех тестов'
+	tesks:
+		- 'qunit:all'
+
