+++ grunt/clean.js
@@ -1,0 +1,5 @@
+module.exports = {
+	all: [
+		// Пусто ;)
+	]
+};
