+++ grunt/qunit.js
@@ -1,0 +1,6 @@
+module.exports = {
+	// Запуск всех тестов
+	all: [
+		'./public_html/tests/index.html'
+	]
+};
