+++ grunt/shell.js
@@ -1,0 +1,10 @@
+module.exports = {
+	// Запуск сервера через скрипт shell'a
+	options: {
+		stdout: true,
+		stderr: true
+	},
+	dev: {
+		command: 'npm start' /* запуск сервера */
+	}
+};
